Add defaultSettings and split plugin config
