+++ iota/docusaurus.config.js
@@ -1,11 +1,11 @@
-const math = require('remark-math');
-const katex = require('rehype-katex');
-
 const path = require('path');
+const contentConfigs = require('./contentPlugins');
+const { create_doc_plugin } = require('../src/utils/config');
 
 module.exports = async () => {
-  const rehype_lodash_template = (await import('rehype-lodash-template'))
-    .default;
+  const contentPlugins = (await contentConfigs()).map((contentConfig) =>
+    create_doc_plugin(contentConfig),
+  );
 
   return {
     title: 'IOTA Wiki',
@@ -213,315 +213,7 @@ module.exports = async () => {
       ],
     ],
     plugins: [
-      //docs
-      [
-        '@docusaurus/plugin-content-docs',
-        {
-          id: 'use',
-          path: path.resolve(__dirname, 'use'),
-          routeBasePath: 'use',
-          sidebarPath: require.resolve('./use/sidebars.ts'),
-
-          // General config
-          editUrl: 'https://github.com/iota-wiki/iota-wiki/edit/main/',
-          remarkPlugins: [
-            require('remark-code-import'),
-            require('remark-import-partial'),
-            [
-              rehype_lodash_template,
-              {
-                values: {
-                  network: 'IOTA',
-                },
-              },
-            ],
-          ],
-          showLastUpdateTime: true,
-        },
-      ],
-      [
-        '@docusaurus/plugin-content-docs',
-        {
-          id: 'learn',
-          path: path.resolve(__dirname, 'learn'),
-          routeBasePath: 'learn',
-          sidebarPath: require.resolve('./learn/sidebars.ts'),
-
-          // General config
-          editUrl: 'https://github.com/iota-wiki/iota-wiki/edit/main/',
-          remarkPlugins: [
-            require('remark-code-import'),
-            require('remark-import-partial'),
-          ],
-          showLastUpdateTime: true,
-        },
-      ],
-      [
-        '@docusaurus/plugin-content-docs',
-        {
-          id: 'develop',
-          path: path.resolve(__dirname, 'develop'),
-          routeBasePath: 'develop',
-          sidebarPath: require.resolve('./develop/sidebars.ts'),
-          docItemComponent: '@theme/ApiItem',
-
-          // General config
-          editUrl: 'https://github.com/iota-wiki/iota-wiki/edit/main/',
-          remarkPlugins: [
-            require('remark-code-import'),
-            require('remark-import-partial'),
-          ],
-          showLastUpdateTime: true,
-        },
-      ],
-      [
-        '@docusaurus/plugin-content-docs',
-        {
-          id: 'iota-chronicle-rs',
-          path: path.resolve(__dirname, 'docs/chronicle/docs'),
-          routeBasePath: 'chronicle',
-          sidebarPath: path.resolve(__dirname, 'docs/chronicle/sidebars.js'),
-          editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-          remarkPlugins: [
-            require('remark-code-import'),
-            require('remark-import-partial'),
-          ],
-          showLastUpdateTime: true,
-          versions: {
-            current: {
-              label: 'IOTA',
-              badge: true,
-            },
-          },
-        },
-      ],
-      [
-        '@docusaurus/plugin-content-docs',
-        {
-          id: 'iota-goshimmer',
-          path: path.resolve(__dirname, 'docs/goshimmer/docs'),
-          routeBasePath: 'goshimmer',
-          sidebarPath: path.resolve(__dirname, 'docs/goshimmer/sidebars.js'),
-          editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-          showLastUpdateTime: true,
-        },
-      ],
-      [
-        '@docusaurus/plugin-content-docs',
-        {
-          id: 'iota-hornet',
-          path: path.resolve(__dirname, 'docs/hornet/docs'),
-          routeBasePath: 'hornet',
-          sidebarPath: path.resolve(__dirname, 'docs/hornet/sidebars.js'),
-          editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-          showLastUpdateTime: true,
-          versions: {
-            current: {
-              label: 'IOTA',
-              badge: true,
-            },
-          },
-        },
-      ],
-      [
-        '@docusaurus/plugin-content-docs',
-        {
-          id: 'iota-integration-services',
-          path: path.resolve(__dirname, 'docs/integration-services/docs'),
-          routeBasePath: 'integration-services',
-          sidebarPath: path.resolve(
-            __dirname,
-            'docs/integration-services/sidebars.js',
-          ),
-          editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-          showLastUpdateTime: true,
-        },
-      ],
-      [
-        '@docusaurus/plugin-content-docs',
-        {
-          id: 'iota-introduction-docs',
-          path: path.resolve(__dirname, 'docs/introduction-docs/docs'),
-          routeBasePath: 'introduction',
-          sidebarPath: path.resolve(
-            __dirname,
-            'docs/introduction-docs/sidebars.js',
-          ),
-          editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-          remarkPlugins: [
-            require('remark-code-import'),
-            require('remark-import-partial'),
-            require('remark-remove-comments'),
-          ],
-          showLastUpdateTime: true,
-          versions: {
-            current: {
-              label: 'IOTA',
-              badge: true,
-            },
-          },
-        },
-      ],
-      [
-        '@docusaurus/plugin-content-docs',
-        {
-          id: 'iota-IOTA-Research-Specifications',
-          path: path.resolve(
-            __dirname,
-            'docs/iota-2.0-research-specifications/',
-          ),
-          routeBasePath: 'IOTA-2.0-Research-Specifications',
-          sidebarPath: path.resolve(
-            __dirname,
-            'docs/iota-2.0-research-specifications/sidebars.js',
-          ),
-          editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-          remarkPlugins: [math],
-          rehypePlugins: [katex],
-          showLastUpdateTime: true,
-          include: ['*.md'],
-        },
-      ],
-      [
-        '@docusaurus/plugin-content-docs',
-        {
-          id: 'iota-streams',
-          path: path.resolve(__dirname, 'docs/streams/docs'),
-          routeBasePath: 'streams',
-          sidebarPath: path.resolve(__dirname, 'docs/streams/sidebars.js'),
-          editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-          showLastUpdateTime: true,
-        },
-      ],
-      [
-        '@docusaurus/plugin-content-docs',
-        {
-          id: 'iota-zebra-tutorials',
-          path: path.resolve(__dirname, 'docs/zebra-iota-edge-sdk/tutorials'),
-          routeBasePath: 'zebra-tutorials',
-          sidebarPath: path.resolve(
-            __dirname,
-            'docs/zebra-iota-edge-sdk/sidebars.js',
-          ),
-          editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-          showLastUpdateTime: true,
-        },
-      ],
-      [
-        '@docusaurus/plugin-content-docs',
-        {
-          id: 'stronghold-rs',
-          path: path.resolve(__dirname, 'docs/stronghold.rs/docs'),
-          routeBasePath: 'stronghold.rs',
-          sidebarPath: path.resolve(
-            __dirname,
-            'docs/stronghold.rs/sidebars.js',
-          ),
-          editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-          remarkPlugins: [
-            require('remark-code-import'),
-            require('remark-import-partial'),
-            require('remark-remove-comments'),
-            math,
-          ],
-          rehypePlugins: [katex],
-          showLastUpdateTime: true,
-        },
-      ],
-      [
-        '@docusaurus/plugin-content-docs',
-        {
-          id: 'wallet-rs',
-          path: path.resolve(__dirname, 'docs/wallet.rs/docs'),
-          routeBasePath: 'wallet.rs',
-          sidebarPath: path.resolve(__dirname, 'docs/wallet.rs/sidebars.js'),
-          editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-          remarkPlugins: [
-            require('remark-code-import'),
-            require('remark-import-partial'),
-          ],
-          showLastUpdateTime: true,
-          versions: {
-            current: {
-              label: 'IOTA',
-              badge: true,
-            },
-          },
-        },
-      ],
-      [
-        '@docusaurus/plugin-content-docs',
-        {
-          id: 'iota-rs',
-          path: path.resolve(__dirname, 'docs/iota.rs/docs'),
-          routeBasePath: 'iota.rs',
-          sidebarPath: path.resolve(__dirname, 'docs/iota.rs/sidebars.js'),
-          editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-          remarkPlugins: [
-            require('remark-code-import'),
-            require('remark-import-partial'),
-          ],
-          showLastUpdateTime: true,
-          versions: {
-            current: {
-              label: 'IOTA',
-              badge: true,
-            },
-          },
-        },
-      ],
-      [
-        '@docusaurus/plugin-content-docs',
-        {
-          id: 'identity-rs-0-5',
-          path: path.resolve(__dirname, 'docs/identity.rs/v0.5.0/docs'),
-          routeBasePath: 'identity.rs',
-          sidebarPath: path.resolve(
-            __dirname,
-            'docs/identity.rs/v0.5.0/sidebars.js',
-          ),
-          editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-          remarkPlugins: [
-            require('remark-code-import'),
-            require('remark-import-partial'),
-            require('remark-remove-comments'),
-          ],
-          showLastUpdateTime: true,
-          versions: {
-            current: {
-              label: '0.5',
-              path: '0.5',
-              badge: true,
-            },
-          },
-        },
-      ],
-      [
-        '@docusaurus/plugin-content-docs',
-        {
-          id: 'identity-rs-0-6',
-          path: path.resolve(__dirname, 'docs/identity.rs/v0.6.0/docs'),
-          routeBasePath: 'identity.rs',
-          sidebarPath: path.resolve(
-            __dirname,
-            'docs/identity.rs/v0.6.0/sidebars.js',
-          ),
-          editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-          remarkPlugins: [
-            require('remark-code-import'),
-            require('remark-import-partial'),
-            require('remark-remove-comments'),
-          ],
-          showLastUpdateTime: true,
-          versions: {
-            current: {
-              label: '0.6',
-              badge: true,
-            },
-          },
-        },
-      ],
-      //plugins
+      ...contentPlugins,
       [
         'docusaurus-plugin-openapi-docs',
         {
