+++ next/docusaurus.config.js
@@ -1,6 +1,10 @@
 const path = require('path');
-const math = require('remark-math');
-const katex = require('rehype-katex');
+const contentConfigs = require('./contentPlugins');
+const { create_doc_plugin } = require('../src/utils/config');
+
+const contentPlugins = contentConfigs.map((contentConfig) =>
+  create_doc_plugin(contentConfig),
+);
 
 module.exports = {
   title: 'Danger Zone',
@@ -198,371 +202,7 @@ module.exports = {
     ],
   ],
   plugins: [
-    //docs
-    [
-      '@docusaurus/plugin-content-docs',
-      {
-        id: 'use',
-        path: path.resolve(__dirname, 'use'),
-        routeBasePath: 'use',
-        sidebarPath: require.resolve('./use/sidebars.ts'),
-
-        // General config
-        editUrl: 'https://github.com/iota-wiki/iota-wiki/edit/main/',
-        remarkPlugins: [
-          require('remark-code-import'),
-          require('remark-import-partial'),
-        ],
-        showLastUpdateTime: true,
-      },
-    ],
-    [
-      '@docusaurus/plugin-content-docs',
-      {
-        id: 'learn',
-        path: path.resolve(__dirname, 'learn'),
-        routeBasePath: 'learn',
-        sidebarPath: require.resolve('./learn/sidebars.ts'),
-
-        // General config
-        editUrl: 'https://github.com/iota-wiki/iota-wiki/edit/main/',
-        remarkPlugins: [
-          require('remark-code-import'),
-          require('remark-import-partial'),
-        ],
-        showLastUpdateTime: true,
-      },
-    ],
-    [
-      '@docusaurus/plugin-content-docs',
-      {
-        id: 'develop',
-        path: path.resolve(__dirname, 'develop'),
-        routeBasePath: 'develop',
-        sidebarPath: require.resolve('./develop/sidebars.ts'),
-        docItemComponent: '@theme/ApiItem',
-
-        // General config
-        editUrl: 'https://github.com/iota-wiki/iota-wiki/edit/main/',
-        remarkPlugins: [
-          require('remark-code-import'),
-          require('remark-import-partial'),
-        ],
-        showLastUpdateTime: true,
-      },
-    ],
-    [
-      '@docusaurus/plugin-content-docs',
-      {
-        id: 'iota-sdk',
-        path: path.resolve(__dirname, 'docs/iota-sdk/docs'),
-        routeBasePath: 'iota-sdk',
-        sidebarPath: require.resolve('./docs/iota-sdk/sidebars.js'),
-        editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-        remarkPlugins: [
-          require('remark-code-import'),
-          require('remark-import-partial'),
-        ],
-        showLastUpdateTime: true,
-        //overriding default exclude array to include the python api's classes with _ at the beginning
-        //but still exclude any _admonitions
-        exclude: [
-          // '**/_*.{js,jsx,ts,tsx,md}',
-          // '**/_*/**',
-          '**/*.test.{js,jsx,ts,tsx}',
-          '**/__tests__/**',
-          '**/_admonitions/_**',
-        ],
-      },
-    ],
-    [
-      '@docusaurus/plugin-content-docs',
-      {
-        id: 'cli-wallet',
-        path: path.resolve(__dirname, 'docs/cli-wallet/docs'),
-        routeBasePath: 'cli-wallet',
-        sidebarPath: require.resolve('./docs/cli-wallet/sidebars.js'),
-        editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-        showLastUpdateTime: true,
-      },
-    ],
-
-    [
-      '@docusaurus/plugin-content-docs',
-      {
-        id: 'hornet',
-        path: path.resolve(__dirname, 'docs/hornet/docs'),
-        routeBasePath: 'hornet',
-        sidebarPath: path.resolve(__dirname, 'docs/hornet/sidebars.js'),
-        editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-        showLastUpdateTime: true,
-      },
-    ],
-    [
-      '@docusaurus/plugin-content-docs',
-      {
-        id: 'iota-goshimmer',
-        path: path.resolve(__dirname, 'docs/goshimmer/docs'),
-        routeBasePath: 'goshimmer',
-        sidebarPath: path.resolve(__dirname, 'docs/goshimmer/sidebars.js'),
-        editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-        showLastUpdateTime: true,
-      },
-    ],
-    [
-      '@docusaurus/plugin-content-docs',
-      {
-        id: 'introduction-docs',
-        path: path.resolve(__dirname, 'docs/introduction-docs/docs'),
-        routeBasePath: 'introduction',
-        sidebarPath: path.resolve(
-          __dirname,
-          'docs/introduction-docs/sidebars.js',
-        ),
-        editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-        remarkPlugins: [
-          require('remark-code-import'),
-          require('remark-import-partial'),
-          require('remark-remove-comments'),
-        ],
-        showLastUpdateTime: true,
-      },
-    ],
-    [
-      '@docusaurus/plugin-content-docs',
-      {
-        id: 'iota-IOTA-Research-Specifications',
-        path: path.resolve(__dirname, 'docs/iota-2.0-research-specifications/'),
-        routeBasePath: 'IOTA-2.0-Research-Specifications',
-        sidebarPath: path.resolve(
-          __dirname,
-          'docs/iota-2.0-research-specifications/sidebars.js',
-        ),
-        editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-        remarkPlugins: [math],
-        rehypePlugins: [katex],
-        showLastUpdateTime: true,
-        include: ['*.md'],
-      },
-    ],
-    [
-      '@docusaurus/plugin-content-docs',
-      {
-        id: 'wasp',
-        path: path.resolve(__dirname, 'docs/wasp/docs'),
-        routeBasePath: 'smart-contracts',
-        sidebarPath: path.resolve(__dirname, 'docs/wasp/sidebars.js'),
-        editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-        remarkPlugins: [
-          require('remark-code-import'),
-          require('remark-import-partial'),
-          require('remark-remove-comments'),
-        ],
-        showLastUpdateTime: true,
-      },
-    ],
-    [
-      '@docusaurus/plugin-content-docs',
-      {
-        id: 'iota-chronicle-rs',
-        path: path.resolve(__dirname, 'docs/chronicle/docs'),
-        routeBasePath: 'chronicle',
-        sidebarPath: path.resolve(__dirname, 'docs/chronicle/sidebars.js'),
-        editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-        remarkPlugins: [
-          require('remark-code-import'),
-          require('remark-import-partial'),
-        ],
-        showLastUpdateTime: true,
-        versions: {
-          current: {
-            label: 'IOTA',
-            badge: true,
-          },
-        },
-      },
-    ],
-    [
-      '@docusaurus/plugin-content-docs',
-      {
-        id: 'identity-rs',
-        path: path.resolve(__dirname, 'docs/identity.rs/0.7-alpha/docs'),
-        routeBasePath: 'identity.rs',
-        sidebarPath: path.resolve(
-          __dirname,
-          'docs/identity.rs/0.7-alpha/sidebars.js',
-        ),
-        editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-        remarkPlugins: [
-          require('remark-code-import'),
-          require('remark-import-partial'),
-          require('remark-remove-comments'),
-        ],
-        showLastUpdateTime: true,
-        versions: {
-          current: {
-            label: '0.7',
-            badge: true,
-          },
-        },
-      },
-    ],
-    [
-      '@docusaurus/plugin-content-docs',
-      {
-        id: 'inx-api-core-v0',
-        path: path.resolve(__dirname, 'docs/inx-api-core-v0/docs'),
-        routeBasePath: 'inx-api-core-v0',
-        sidebarPath: path.resolve(
-          __dirname,
-          'docs/inx-api-core-v0/sidebars.js',
-        ),
-        editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-        showLastUpdateTime: true,
-      },
-    ],
-
-    [
-      '@docusaurus/plugin-content-docs',
-      {
-        id: 'inx-api-core-v1',
-        path: path.resolve(__dirname, 'docs/inx-api-core-v1/docs'),
-        routeBasePath: 'inx-api-core-v1',
-        sidebarPath: path.resolve(
-          __dirname,
-          'docs/inx-api-core-v1/sidebars.js',
-        ),
-        editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-        showLastUpdateTime: true,
-      },
-    ],
-    [
-      '@docusaurus/plugin-content-docs',
-      {
-        id: 'inx-coordinator',
-        path: path.resolve(__dirname, 'docs/inx-coordinator/docs'),
-        routeBasePath: 'inx-coordinator',
-        sidebarPath: path.resolve(
-          __dirname,
-          'docs/inx-coordinator/sidebars.js',
-        ),
-        editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-        showLastUpdateTime: true,
-      },
-    ],
-    [
-      '@docusaurus/plugin-content-docs',
-      {
-        id: 'inx-dashboard',
-        path: path.resolve(__dirname, 'docs/inx-dashboard/docs'),
-        routeBasePath: 'inx-dashboard',
-        sidebarPath: path.resolve(__dirname, 'docs/inx-dashboard/sidebars.js'),
-        editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-        showLastUpdateTime: true,
-      },
-    ],
-    [
-      '@docusaurus/plugin-content-docs',
-      {
-        id: 'inx-faucet',
-        path: path.resolve(__dirname, 'docs/inx-faucet/docs'),
-        routeBasePath: 'inx-faucet',
-        sidebarPath: path.resolve(__dirname, 'docs/inx-faucet/sidebars.js'),
-        editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-        showLastUpdateTime: true,
-      },
-    ],
-    [
-      '@docusaurus/plugin-content-docs',
-      {
-        id: 'inx-indexer',
-        path: path.resolve(__dirname, 'docs/inx-indexer/docs'),
-        routeBasePath: 'inx-indexer',
-        sidebarPath: path.resolve(__dirname, 'docs/inx-indexer/sidebars.js'),
-        editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-        showLastUpdateTime: true,
-      },
-    ],
-    [
-      '@docusaurus/plugin-content-docs',
-      {
-        id: 'inx-irc-metadata',
-        path: path.resolve(__dirname, 'docs/inx-irc-metadata/docs'),
-        routeBasePath: 'inx-irc-metadata',
-        sidebarPath: path.resolve(
-          __dirname,
-          'docs/inx-irc-metadata/sidebars.js',
-        ),
-        editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-        showLastUpdateTime: true,
-      },
-    ],
-    [
-      '@docusaurus/plugin-content-docs',
-      {
-        id: 'inx-mqtt',
-        path: path.resolve(__dirname, 'docs/inx-mqtt/docs'),
-        routeBasePath: 'inx-mqtt',
-        sidebarPath: path.resolve(__dirname, 'docs/inx-mqtt/sidebars.js'),
-        editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-        showLastUpdateTime: true,
-      },
-    ],
-    [
-      '@docusaurus/plugin-content-docs',
-      {
-        id: 'inx-participation',
-        path: path.resolve(__dirname, 'docs/inx-participation/docs'),
-        routeBasePath: 'inx-participation',
-        sidebarPath: path.resolve(
-          __dirname,
-          'docs/inx-participation/sidebars.js',
-        ),
-        editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-        showLastUpdateTime: true,
-      },
-    ],
-    [
-      '@docusaurus/plugin-content-docs',
-      {
-        id: 'inx-poi',
-        path: path.resolve(__dirname, 'docs/inx-poi/docs'),
-        routeBasePath: 'inx-poi',
-        sidebarPath: path.resolve(__dirname, 'docs/inx-poi/sidebars.js'),
-        editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-        showLastUpdateTime: true,
-      },
-    ],
-    [
-      '@docusaurus/plugin-content-docs',
-      {
-        id: 'inx-spammer',
-        path: path.resolve(__dirname, 'docs/inx-spammer/docs'),
-        routeBasePath: 'inx-spammer',
-        sidebarPath: path.resolve(__dirname, 'docs/inx-spammer/sidebars.js'),
-        editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-        showLastUpdateTime: true,
-      },
-    ],
-    [
-      '@docusaurus/plugin-content-docs',
-      {
-        id: 'stronghold-rs',
-        path: path.resolve(__dirname, 'docs/stronghold.rs/docs'),
-        routeBasePath: 'stronghold.rs',
-        sidebarPath: path.resolve(__dirname, 'docs/stronghold.rs/sidebars.js'),
-        editUrl: 'https://github.com/iotaledger/iota-wiki/edit/main/',
-        remarkPlugins: [
-          require('remark-code-import'),
-          require('remark-import-partial'),
-          require('remark-remove-comments'),
-          math,
-        ],
-        showLastUpdateTime: true,
-        rehypePlugins: [katex],
-      },
-    ],
-    //plugins
+    ...contentPlugins,
     [
       'docusaurus-plugin-openapi-docs',
       {
